fix: fixed HDF5Array availability

+++ tests/testthat/test_aggregate.R
@@ -1,4 +1,3 @@
-library(HDF5Array)
 
 test_that("aggregation: medianPolish", {
     ## numerical example taken from ?stats::medpolish
@@ -65,9 +64,6 @@ test_that("aggregation: aggregate_by_vector", {
                                       "X53", "X55"),
                                     c("iTRAQ4.114", "iTRAQ4.115",
                                       "iTRAQ4.116", "iTRAQ4.117")))
-    ## Data in HDF5
-    tmpf <- paste0(tempdir(), "/xhdf5")
-    xhdf5 <- writeHDF5Array(x, tmpf, with.dimnames = TRUE)
     ## Different ways to provide INDEX
     k_char <- c("B", "E", "X", "E", "B", "B", "E")
     k_fact <- factor(k_char)
@@ -95,17 +91,6 @@ test_that("aggregation: aggregate_by_vector", {
                  aggregate_by_vector(x, k_fact, robustSummary)[same_row_names, ])
     expect_equal(x2_robust_expected[same_row_names, ],
                  aggregate_by_vector(x, k_fact2, robustSummary)[same_row_names, ])
-    ## Test HDF5Matrix compatibility
-    hdf5out <- aggregate_by_vector(xhdf5, k_char, robustSummary)
-    expect_true(inherits(hdf5out, "HDF5Matrix"))
-    expect_equal(x2_robust_expected[same_row_names, ],
-                 as.matrix(hdf5out)[same_row_names, ])
-    hdf5out <- aggregate_by_vector(xhdf5, k_fact, robustSummary)
-    expect_equal(x2_robust_expected[same_row_names, ],
-                 as.matrix(hdf5out)[same_row_names, ])
-    hdf5out <- aggregate_by_vector(xhdf5, k_fact2, robustSummary)
-    expect_equal(x2_robust_expected[same_row_names, ],
-                 as.matrix(hdf5out)[same_row_names, ])
     ## aggregate: medianPolish
     x2_medpolish_expected <-
         structure(c(3.36717083720277, 3.63886529932001,
@@ -125,17 +110,6 @@ test_that("aggregation: aggregate_by_vector", {
                  aggregate_by_vector(x, k_fact, medianPolish)[same_row_names, ])
     expect_equal(x2_medpolish_expected[same_row_names, ],
                  aggregate_by_vector(x, k_fact2, medianPolish)[same_row_names, ])
-    ## Test HDF5Matrix compatibility
-    hdf5out <- aggregate_by_vector(xhdf5, k_char, medianPolish)
-    expect_true(inherits(hdf5out, "HDF5Matrix"))
-    expect_equal(x2_medpolish_expected[same_row_names, ],
-                 as.matrix(hdf5out)[same_row_names, ])
-    hdf5out <- aggregate_by_vector(xhdf5, k_fact, medianPolish)
-    expect_equal(x2_medpolish_expected[same_row_names, ],
-                 as.matrix(hdf5out)[same_row_names, ])
-    hdf5out <- aggregate_by_vector(xhdf5, k_fact2, medianPolish)
-    expect_equal(x2_medpolish_expected[same_row_names, ],
-                 as.matrix(hdf5out)[same_row_names, ])
     ## aggregate: sum
     x2_sum_expected <-
         structure(c(9.70155804007753, 9.75542561310774,
@@ -155,17 +129,48 @@ test_that("aggregation: aggregate_by_vector", {
                  aggregate_by_vector(x, k_fact, colSums)[same_row_names, ])
     expect_equal(x2_sum_expected[same_row_names, ],
                  aggregate_by_vector(x, k_fact2, colSums)[same_row_names, ])
+    
     ## Test HDF5Matrix compatibility
-    hdf5out <- aggregate_by_vector(xhdf5, k_char, Matrix::colSums)
-    expect_true(inherits(hdf5out, "HDF5Matrix"))
-    expect_equal(x2_sum_expected[same_row_names, ],
-                 as.matrix(hdf5out)[same_row_names, ])
-    hdf5out <- aggregate_by_vector(xhdf5, k_fact, Matrix::colSums)
-    expect_equal(x2_sum_expected[same_row_names, ],
-                 as.matrix(hdf5out)[same_row_names, ])
-    hdf5out <- aggregate_by_vector(xhdf5, k_fact2, Matrix::colSums)
-    expect_equal(x2_sum_expected[same_row_names, ],
-                 as.matrix(hdf5out)[same_row_names, ])
+    if (requireNamespace("HDF5Array", quietly = TRUE)) {
+        ## Data in HDF5
+        tmpf <- paste0(tempdir(), "/xhdf5")
+        xhdf5 <- HDF5Array::writeHDF5Array(x, tmpf, with.dimnames = TRUE)
+        ## robustSummary
+        hdf5out <- aggregate_by_vector(xhdf5, k_char, robustSummary)
+        expect_true(inherits(hdf5out, "HDF5Matrix"))
+        expect_equal(x2_robust_expected[same_row_names, ],
+                     as.matrix(hdf5out)[same_row_names, ])
+        hdf5out <- aggregate_by_vector(xhdf5, k_fact, robustSummary)
+        expect_equal(x2_robust_expected[same_row_names, ],
+                     as.matrix(hdf5out)[same_row_names, ])
+        hdf5out <- aggregate_by_vector(xhdf5, k_fact2, robustSummary)
+        expect_equal(x2_robust_expected[same_row_names, ],
+                     as.matrix(hdf5out)[same_row_names, ])
+        ## medianPolish
+        hdf5out <- aggregate_by_vector(xhdf5, k_char, medianPolish)
+        expect_true(inherits(hdf5out, "HDF5Matrix"))
+        expect_equal(x2_medpolish_expected[same_row_names, ],
+                     as.matrix(hdf5out)[same_row_names, ])
+        hdf5out <- aggregate_by_vector(xhdf5, k_fact, medianPolish)
+        expect_equal(x2_medpolish_expected[same_row_names, ],
+                     as.matrix(hdf5out)[same_row_names, ])
+        hdf5out <- aggregate_by_vector(xhdf5, k_fact2, medianPolish)
+        expect_equal(x2_medpolish_expected[same_row_names, ],
+                     as.matrix(hdf5out)[same_row_names, ])
+        ## Matrix::colSums
+        hdf5out <- aggregate_by_vector(xhdf5, k_char, Matrix::colSums)
+        expect_true(inherits(hdf5out, "HDF5Matrix"))
+        expect_equal(x2_sum_expected[same_row_names, ],
+                     as.matrix(hdf5out)[same_row_names, ])
+        hdf5out <- aggregate_by_vector(xhdf5, k_fact, Matrix::colSums)
+        expect_equal(x2_sum_expected[same_row_names, ],
+                     as.matrix(hdf5out)[same_row_names, ])
+        hdf5out <- aggregate_by_vector(xhdf5, k_fact2, Matrix::colSums)
+        expect_equal(x2_sum_expected[same_row_names, ],
+                     as.matrix(hdf5out)[same_row_names, ])
+        ## Remove file
+        unlink(tmpf)
+    }
 })
 
 
@@ -174,11 +179,6 @@ test_that("aggregation: colCounts", {
     m <- matrix(c(1, NA, 2, 3, NA, NA, 4, 5, 6),
                 nrow = 3)
     expect_identical(colCounts(m), c(2, 1, 3))
-    ## Test HDF5Matrix compatibility
-    tmpf <- paste0(tempdir(), "/mhdf5")
-    mhdf5 <- writeHDF5Array(m, tmpf, with.dimnames = TRUE)
-    expect_identical(colCounts(mhdf5), c(2, 1, 3))
-    unlink(tmpf)
     ## No NAs
     m <- matrix(rnorm(30), nrow = 3)
     expect_identical(colCounts(m), rep(3, 10))
@@ -196,8 +196,18 @@ test_that("aggregation: colCounts", {
     ## NA and Inf
     m[2,2] <- Inf
     expect_identical(colCounts(m), c(4, rep(5, 4)))
+    
+    ## Test HDF5Matrix compatibility
+    if (requireNamespace("HDF5Array", quietly = TRUE)) {
+        ## Test HDF5Matrix compatibility
+        tmpf <- paste0(tempdir(), "/mhdf5")
+        m <- matrix(c(1, NA, 2, 3, NA, NA, 4, 5, 6),
+                    nrow = 3)
+        mhdf5 <- HDF5Array::writeHDF5Array(m, tmpf, with.dimnames = TRUE)
+        expect_identical(colCounts(mhdf5), c(2, 1, 3))
+        unlink(tmpf)
+    }
 })
-
 
 
 test_that("aggregate_by_matix works", {
@@ -209,9 +219,6 @@ test_that("aggregate_by_matix works", {
                c( 0, 2, 0))
     colnames(x) <- paste0("S", 1:3)
     rownames(x) <- paste0("pep", 1:5)
-    ## Data in HDF5
-    tmpf <- paste0(tempdir(), "/mhdf5")
-    xhdf5 <- writeHDF5Array(x, tmpf, with.dimnames = TRUE)
     ## aggregation index
     k <- paste0("Prot", rep(1:2, c(2, 3)))
     ## adjacency matrix
@@ -221,21 +228,29 @@ test_that("aggregate_by_matix works", {
     rownames(adj) <- rownames(x)
     colnames(adj) <- unique(k)
     ## aggregation by sum
-    av <- aggregate_by_vector(x, k, colSums)
-    am <- aggregate_by_matrix(x, adj, colSumsMat)
-    amhdf5 <- aggregate_by_matrix(xhdf5, adj, colSumsMat)
-    expect_identical(av, am)
-    expect_true(inherits(amhdf5, "HDF5Matrix"))
-    expect_identical(am, as.matrix(amhdf5))
+    av1 <- aggregate_by_vector(x, k, colSums)
+    am1 <- aggregate_by_matrix(x, adj, colSumsMat)
+    expect_identical(av1, am1)
     ## aggregation by mean
-    av <- aggregate_by_vector(x, k, colMeans)
-    am <- aggregate_by_matrix(x, adj, colMeansMat)
-    amhdf5 <- aggregate_by_matrix(xhdf5, adj, colMeansMat)
-    expect_identical(av, am)
-    expect_true(inherits(amhdf5, "HDF5Matrix"))
-    expect_identical(am, as.matrix(amhdf5))
-    
-    unlink(tmpf)
+    av2 <- aggregate_by_vector(x, k, colMeans)
+    am2 <- aggregate_by_matrix(x, adj, colMeansMat)
+    expect_identical(av2, am2)
+    ## Test HDF5Matrix compatibility
+    if (requireNamespace("HDF5Array", quietly = TRUE)) {
+        ## Data in HDF5
+        tmpf <- paste0(tempdir(), "/mhdf5")
+        xhdf5 <- HDF5Array::writeHDF5Array(x, tmpf, with.dimnames = TRUE)
+        ## colSumsMat
+        amhdf5 <- aggregate_by_matrix(xhdf5, adj, colSumsMat)
+        expect_true(inherits(amhdf5, "HDF5Matrix"))
+        expect_identical(am1, as.matrix(amhdf5))
+        ## colMeansMat
+        amhdf5 <- aggregate_by_matrix(xhdf5, adj, colMeansMat)
+        expect_true(inherits(amhdf5, "HDF5Matrix"))
+        expect_identical(am2, as.matrix(amhdf5))
+        ## Remove file
+        unlink(tmpf)
+    }
 })
 
 
@@ -249,9 +264,6 @@ test_that("aggregate_by_matix works with NAs", {
     colnames(x) <- paste0("S", 1:3)
     rownames(x) <- paste0("pep", 1:5)
     x[1, 1] <- x[4, 2] <- NA
-    ## Data in HDF5
-    tmpf <- paste0(tempdir(), "/mhdf5")
-    xhdf5 <- writeHDF5Array(x, tmpf, with.dimnames = TRUE)
     ## aggregation index
     k <- paste0("Prot", rep(1:2, c(2, 3)))
     ## adjacency matrix
@@ -261,33 +273,40 @@ test_that("aggregate_by_matix works with NAs", {
     rownames(adj) <- rownames(x)
     colnames(adj) <- unique(k)
     ## aggregation by sum, na.rm = FALSE
-    av <- aggregate_by_vector(x, k, colSums)
-    am <- aggregate_by_matrix(x, adj, colSumsMat)
-    amhdf5 <- aggregate_by_matrix(xhdf5, adj, colSumsMat)
-    expect_identical(av, am)
-    expect_identical(am, as.matrix(amhdf5))
+    av1 <- aggregate_by_vector(x, k, colSums)
+    am1 <- aggregate_by_matrix(x, adj, colSumsMat)
+    expect_identical(av1, am1)
     ## aggregation by mean, na.rm = FALSE
-    av <- aggregate_by_vector(x, k, colMeans)
-    am <- aggregate_by_matrix(x, adj, colMeansMat)
-    amhdf5 <- aggregate_by_matrix(xhdf5, adj, colMeansMat)
-    expect_true(inherits(amhdf5, "HDF5Matrix"))
-    expect_identical(av, am)
-    expect_identical(am, as.matrix(amhdf5))
+    av2 <- aggregate_by_vector(x, k, colMeans)
+    am2 <- aggregate_by_matrix(x, adj, colMeansMat)
+    expect_identical(av2, am2)
     ## aggregation by sum, na.rm = TRUE
-    av <- aggregate_by_vector(x, k, colSums, na.rm = TRUE)
-    am <- aggregate_by_matrix(x, adj, colSumsMat, na.rm = TRUE)
-    amhdf5 <- aggregate_by_matrix(xhdf5, adj, colSumsMat, na.rm = TRUE)
-    expect_true(inherits(amhdf5, "HDF5Matrix"))
-    expect_identical(av, am)
-    expect_identical(am, as.matrix(amhdf5))
+    av3 <- aggregate_by_vector(x, k, colSums, na.rm = TRUE)
+    am3 <- aggregate_by_matrix(x, adj, colSumsMat, na.rm = TRUE)
+    expect_identical(av3, am3)
     ## aggregation by mean, na.rm = TRUE
-    av <- aggregate_by_vector(x, k, colMeans, na.rm = TRUE)
-    am <- aggregate_by_matrix(x, adj, colMeansMat, na.rm = TRUE)
-    amhdf5 <- aggregate_by_matrix(xhdf5, adj, colMeansMat, na.rm = TRUE)
-    expect_true(inherits(amhdf5, "HDF5Matrix"))
-    expect_identical(av, am)
-    expect_identical(am, as.matrix(amhdf5))
-    
-    unlink(amhdf5)
+    av4 <- aggregate_by_vector(x, k, colMeans, na.rm = TRUE)
+    am4 <- aggregate_by_matrix(x, adj, colMeansMat, na.rm = TRUE)
+    expect_identical(av4, am4)
+    ## Test HDF5Matrix compatibility
+    if (requireNamespace("HDF5Array", quietly = TRUE)) {
+        ## Data in HDF5
+        tmpf <- paste0(tempdir(), "/mhdf5")
+        xhdf5 <- HDF5Array::writeHDF5Array(x, tmpf, with.dimnames = TRUE)
+        ## aggregation by sum, na.rm = FALSE
+        amhdf5 <- aggregate_by_matrix(xhdf5, adj, colSumsMat)
+        expect_identical(am1, as.matrix(amhdf5))
+        ## aggregation by mean, na.rm = FALSE
+        amhdf5 <- aggregate_by_matrix(xhdf5, adj, colMeansMat)
+        expect_identical(am2, as.matrix(amhdf5))
+        ## aggregation by sum, na.rm = TRUE
+        amhdf5 <- aggregate_by_matrix(xhdf5, adj, colSumsMat, na.rm = TRUE)
+        expect_identical(am3, as.matrix(amhdf5))
+        ## aggregation by mean, na.rm = TRUE
+        amhdf5 <- aggregate_by_matrix(xhdf5, adj, colMeansMat, na.rm = TRUE)
+        expect_identical(am4, as.matrix(amhdf5))
+        ## Remove file
+        unlink(amhdf5)
+    }
 })
